Add policynv support
This commit introduces support for policynv, which enables the
validation of sealed tokens against a known good state in an NV index.
This feature enhances protection against reply or downgrade attacks that
may occur with outdated tokens. In practice, a monotonic counter or
timestamp is used to describe the token's generation, which can be
entrusted or distrusted by accounting authorized TPM NV indices.

In contrast, the "fdectl regenerate-key" process is used for key
rotation and revocation of all TPM2 tokens. However, this procedure
involves manipulating on-disk key slots, and any failure during the
process can result in a lockout. Therefore, tpm2 policynv may be a
preferred method, as it can maintain the on-disk LUKS2 header's
integrity.

Signed-off-by: Michael Chang <mchang@suse.com>